Tn7_Transposition/requirements.txt
biopython==1.78
cycler==0.10.0
et-xmlfile==1.1.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
openpyxl==3.0.9
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
xlrd==2.0.1
